Add an HttpHandler that can handle simultaneous requests
Asp.NET will queue requests (for a single session) unless the HttpHandler they hit implements IReadOnlySessionState.
Server Sent Events hold a request open semi-permanently, so any subsequent requests would just sit in the queue waiting
for the SSE request to finish.

I've added an new handler just for this, which feels pretty heavy handed. I'm also not thrilled about overriding .ashx, but "when in rome".

If anyone has a better idea for this I'd be thrilled to hear it!